Fixing sorting of Project Timeline list
